chore: convert versions, deployments, and rollback to defineCommand
