+++ packages/wrangler/src/versions/deployments/index.ts
@@ -1,37 +1,11 @@
-import {
-	versionsDeploymentsListHandler,
-	versionsDeploymentsListOptions,
-} from "./list";
-import {
-	versionsDeploymentsStatusHandler,
-	versionsDeploymentsStatusOptions,
-} from "./status";
-import {
-	versionsDeploymentsViewHandler,
-	versionsDeploymentsViewOptions,
-} from "./view";
-import type { CommonYargsArgv } from "../../yargs-types";
+import { defineNamespace } from "../../core";
 
-export default function registerVersionsDeploymentsSubcommands(
-	versionDeploymentsYargs: CommonYargsArgv
-) {
-	versionDeploymentsYargs
-		.command(
-			"list",
-			"Displays the 10 most recent deployments of your Worker",
-			versionsDeploymentsListOptions,
-			versionsDeploymentsListHandler
-		)
-		.command(
-			"status",
-			"View the current state of your production",
-			versionsDeploymentsStatusOptions,
-			versionsDeploymentsStatusHandler
-		)
-		.command(
-			"view [deployment-id]",
-			false,
-			versionsDeploymentsViewOptions,
-			versionsDeploymentsViewHandler
-		);
-}
+defineNamespace({
+	command: "wrangler deployments",
+	metadata: {
+		description:
+			"🚢 List and view the current and past deployments for your Worker",
+		status: "open-beta",
+		owner: "Workers: Authoring and Testing",
+	},
+});
